chore: rename nevigation to navigation
